Server independed
Подправил баги в сервере которые возникали из-за того, что переменные
координат и скорости были не того типа и залил ему все дллки

+++ Server/nova-server/ship.h
@@ -19,10 +19,10 @@ class CShellPosition
 {
 public:
     QPoint *pos;
-    qint64 direction;
-    qint64 speed;
-    qint64 image_angle;
-    qint64 rot_speed;
+    qint32 direction;
+    qint32 speed;
+    qint32 image_angle;
+    qint32 rot_speed;
 public:
     CShellPosition();
 };
